pyproject.toml
[project]
name = "mcp-standard-bots"
version = "0.1.0"
description = "FastMCP Server implementation for Standard Bots routine editor"
authors = [
    {name = "Arjun", email = "arjun@standardbots.com"},
]
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "mcp>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.1.0",
]

[project.scripts]
mcp-standard-bots = "mcp_standard_bots.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/sub-arjun/mcp-standard-bots"
Repository = "https://github.com/sub-arjun/mcp-standard-bots.git"
Documentation = "https://github.com/sub-arjun/mcp-standard-bots#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.isort]
known-first-party = ["mcp_standard_bots"]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=mcp_standard_bots"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
asyncio_mode = "auto" 